==> Notebooks_Diabetes_Prediction/02_Data_Cleaning.ipynb <==
# Project: Predicting Diabetes Onset FHIR EHR Data
### Notebook: 02_Data_Cleaning.ipynb
### Purpose: Clean raw data

import pandas as pd
import numpy as np
from typing import List, Tuple
from pathlib import Path

# Helper function to load data
def load_data(file_names: List[str], path: str) -> List[pd.DataFrame]:
    dfs = []
    base_path = Path(path)
    
    for name in file_names:
        file_path = base_path / name

        try:
            df = pd.read_csv(file_path)
            dfs.append(df)
        except fileNotFoundError:
            print(f"[WARNING] File not found at: {file_path}")
        except pd.errors.EmptyDataError:
            print(f"[WARNING] File empty: {file_path}")
        except Exception as e:
            print(f"Failed to load data at {file_path}:{e}")
    return dfs

# Load data
file_names = ['Conditions_Raw.csv',
             'Patients_Raw.csv',
             'Encounters_Raw.csv',
             'Observations_Raw.csv']
base_path = '/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Raw'
dfs = load_data(file_names, path=base_path)


df_conditions_raw = dfs[0]
df_patients_raw = dfs[1]
df_encounters_raw = dfs[2]
df_observations_raw = dfs[3]

## Preliminary Data Validation of Response Variable (Diabetes)

# check data frame shapes
print(
    f"Patients: {df_patients_raw.shape}\n" 
    f"Observations: {df_observations_raw.shape}\n" 
    f"Encounters: {df_encounters_raw.shape}\n" 
    f"Conditions: {df_conditions_raw.shape}")

# Create a flag for prediabetic patients to do initial validation
# confirming if we have enough prediabetic and diabetic patients

df_conditions_raw["is_prediabetes"] = df_conditions_raw["name"].str.lower().str.contains("prediabetes")

# Create a flag for diabetic patients to do initial validation
# confirming if we have enough prediabetic and diabetic patients

df_conditions_raw["is_diabetes"] = df_conditions_raw["name"].str.lower().str.contains("diabetes")

# count prediabetic patients
n_prediabetic_pts = df_conditions_raw[df_conditions_raw["is_prediabetes"]]["patient_id"].nunique()
n_prediabetic_pts

# count diabetic patients
n_diabetic_pts = df_conditions_raw[df_conditions_raw["is_diabetes"]]["patient_id"].nunique()
n_diabetic_pts

# convert onset column into datetime object
df_conditions_raw['onset'] = pd.to_datetime(df_conditions_raw['onset'], errors='coerce')

# find all pre and diabetic patients,
# merge rows to find how many prediabetic pts converting to diabetic
pre = df_conditions_raw[df_conditions_raw['is_prediabetes']]
dm = df_conditions_raw[df_conditions_raw['is_diabetes']]

# merge
merged = pre.merge(
    dm,
    on='patient_id',
    suffixes=('_pre','_dm')
)


# find patients progressing from prediabetes to diabetes

progressors = merged[merged['onset_pre'] < merged['onset_dm']]
progressors.head()

progressors.shape

## Data Cleaning

# copy data to preserve raw data

df_patients = df_patients_raw.copy()
df_obs = df_observations_raw.copy()
df_enc = df_encounters_raw.copy()
df_cond = df_conditions_raw.copy()

#### Structural Checks

# Helper function to inspect dataframe structure

def inspect_structure(df, name: str, n: int = 5) -> None:
    """inspect structure, columns, data types, and first n rows
    of dataframe

    Parameters:
    df: dataframe
    name: name of dataframe
    n: number of first few rows in dataframe
    """
    print(f"\n=== {name} ===")
    print(f"\nShape: {df.shape}")
    print("\nColumns: ")
    print(list(df.columns))
    print("\nDtypes: ")
    print(df.dtypes)
    print(f"\nPreview (first {n} rows):")
    print(df.head(n))

# Wrapper function to check structure of multiple dataframes at once

def run_structural_checks(dfs: dict) -> None:
    for name, df in dfs.items():
        inspect_structure(df, name)

# dictionary holding all dataframes and their names

dfs = {
    "Patients": df_patients,
    "Observations": df_obs,
    "Encounters": df_enc,
    "Conditions": df_cond
}

# inspect all dataframe structures

run_structural_checks(dfs)

# ensure no NA patient ids and all present in df_patients

assert df_obs['patient_id'].notna().all()
assert df_enc['patient_id'].notna().all()
assert df_cond['patient_id'].notna().all()

assert df_obs['patient_id'].isin(df_patients['patient_id']).all()

invalid_obs = df_obs.loc[
    ~df_obs['patient_id'].isin(df_patients['patient_id'])
]
len(invalid_obs)

df_patients.shape, df_obs.shape

df_patients.columns, df_obs.columns

df_patients.head(), df_obs.head(), df_enc.head(), df_cond.head()

# helper function to format patient_id in all dataframes exactly as seen in df_patients

def strip_urn_uuid(x):
    """remove urn:uuid: prefix from patient_id
    from all dataframes
    """
    
    if pd.isna(x):
        return None
    x = str(x)
    return x.replace("urn:uuid:", "")

# modify patient_id format in all dataframes to match patient_id in df_patients

for df in [df_obs, df_enc, df_cond]:
    df['patient_id'] = df['patient_id'].apply(strip_urn_uuid)

# validate all patient_ids actoss all dataframes match with those in df_patients

assert df_obs['patient_id'].isin(df_patients['patient_id']).all()
assert df_enc['patient_id'].isin(df_patients['patient_id']).all()
assert df_cond['patient_id'].isin(df_patients['patient_id']).all()

#### Data Types and Conversions

# df_obs time column converted to datetime

df_obs['time'] = pd.to_datetime(df_obs['time'], utc=True)
df_obs.dtypes

for col in ['start','end']:
    df_enc[col] = pd.to_datetime(df_enc[col], utc=True)
df_enc.dtypes

for col in ['onset', 'recorded_date']:
    df_cond[col] = pd.to_datetime(df_cond[col], utc=True)
df_cond.dtypes

#### Missing Data

df_obs.isna().mean().sort_values(ascending=False)

(
    df_obs[df_obs['value'].isna()]
    .groupby('name')
    .size()
    .sort_values(ascending=False)
    .head(15)
)
    

#filter rows with numeric data only

df_obs_numeric = df_obs.loc[df_obs['value'].notna() & df_obs['unit'].notna()]
df_obs_numeric.isna().mean()

df_enc.isna().mean().sort_values(ascending=False)

df_cond.isna().mean().sort_values(ascending=False)

# Helper function to find duplicate rows

def find_duplicates(dfs: dict):
    for name, df in dfs.items():
        print(name)
        print(f"Duplicates Count: {df.duplicated().sum()}")

# find duplciates

dfs_updated =  {
    "Patients": df_patients,
    "Observations Numeric": df_obs_numeric,
    "Encounters": df_enc,
    "Conditions": df_cond
}
find_duplicates(dfs_updated)

# inspect observations df for duplicates

subset_cols = ['patient_id','code','time']
print(f"Duplicate observations count: {df_obs_numeric.duplicated(subset=subset_cols,keep=False).sum()}")

# deduplicate observation df based on subset columns

deduped = df_obs_numeric.drop_duplicates(subset=subset_cols)
deduped.duplicated().sum()

# assign deduplicated df back to numeric observation df
df_obs_numeric = deduped

# rerun duplicates check after cleaning

find_duplicates(dfs_updated)

# validate deduplication

# dictionary of deduplicated dataframes
dfs_deduped =  {
    "Patients": df_patients,
    "Observations Numeric": df_obs_numeric,
    "Encounters": df_enc,
    "Conditions": df_cond
}

find_duplicates(dfs_deduped)

#### Semantic Checks

# check for kinds of labs present in observations df
# ideally want vitals, HbA1c, BMI, random glucose, BUN/Cr, filtration rates

df_obs_numeric['name'].value_counts().head(30)

# conditions
df_cond[(df_cond['is_prediabetes'] & df_cond['is_diabetes'])].head(10)

df_enc['class'].value_counts().head(10)

#### Dates Conversion Validation

df_patients['birth_date'] = pd.to_datetime(df_patients['birth_date'], utc=True)

df_patients.dtypes, df_obs_numeric.dtypes, df_enc.dtypes, df_cond.dtypes

## Cohort Segmentation for Modeling

#### Define Cohort
##### We are interested in the adult population so all patients must be 18 years or older.

df_patients.head(2)

# add age column for patients
reference_date = pd.Timestamp.now(tz='UTC')
df_patients['age'] = (reference_date - df_patients['birth_date']).dt.days // 365
df_patients.head()

# filter all patients >= 18 years

df_patients_adult = df_patients.loc[df_patients['age'] >= 18].copy()

# helper function to filter to adult patients
def filter_to_adult(df, adult_ids, id_col='patient_id'):
    """
    filter all patients in dataframe that are present in adult_ids
    """
    return df.loc[df[id_col].isin(adult_ids)].copy()

# filter all dataframes for adult patient only

adult_pt_ids = df_patients_adult['patient_id']

df_obs_adult = filter_to_adult(df_obs_numeric, adult_pt_ids)
df_enc_adult = filter_to_adult(df_enc, adult_pt_ids)
df_cond_adult = filter_to_adult(df_cond, adult_pt_ids)

# check dimensions after filtering

for df in [df_patients_adult,
          df_obs_adult,
          df_enc_adult,
          df_cond_adult]:
    print(df.shape)

## Outcome: Diabetes onset
### Definition:
#### First occurrence of Type 2 Diabetes diagnosis OR
#### First occurrence of diabetes-related complication

### Cases Cohort

# Cases cohort: all patients with diabetes
# filter first date of diabetes diagnosis

df_dm_events = df_cond_adult.loc[df_cond_adult['is_diabetes']].copy()
df_dm_onset = (
    df_dm_events
    .groupby('patient_id')['onset']
    .min()
    .reset_index()
    .rename(columns={'onset':'diabetes_onset_date'})
)

print(
    f"\nDiabetes Cases df Shape: {df_dm_onset.shape}\n"
    f"\nFirst few rows:\n"
    f"\n{df_dm_onset.head()}")

### Controls Cohort

df_last_seen = (
    df_enc_adult
    .groupby('patient_id')['start']
    .max()
    .reset_index()
    .rename(columns={'start':'last_encounter_date'})
)

df_cohort = df_patients_adult[['patient_id']].copy()
df_cohort = df_cohort.merge(df_dm_onset, on='patient_id', how='left')
df_cohort = df_cohort.merge(df_last_seen, on='patient_id', how='left')

df_cohort.head()

labels = df_cohort['diabetes_onset_date'].notna().astype(int)
labels.head(2)

# create labels column
df_cohort['label'] = labels

# index_date will be the date we predict from, either 1 year prior
# from dm diagnosis for dm patients or
# 1 year prior to last encounter date for patients
# who never develop dm

df_cohort['index_date'] = np.where(
    df_cohort['label'] == 1,
    df_cohort['diabetes_onset_date'] - pd.Timedelta(days=365),
    df_cohort['last_encounter_date'] - pd.Timedelta(days=365)
)

df_cohort.head()


# develop features dataframe
# merge observaitons df with cohort df on patient_id

df_obs_features = df_obs_adult.merge(
    df_cohort[['patient_id', 'index_date']],
    on='patient_id'
    )
df_obs_features.head()

# filter observations that are taken 1 year prior to the index_date for each patient
df_obs_features = df_obs_features[
    (df_obs_features['time'] < df_obs_features['index_date']) &
    (df_obs_features['time'] >= df_obs_features['index_date'] - pd.Timedelta(days=365))
]

df_obs_features.head()

# identify vounts of LOINC lab codes
df_obs_features[['code']].value_counts().head(20)


feature_codes = {
    '29463-7',   # weight
    '8302-2',    # height
    '39156-5',   # BMI
    '8867-4',    # heart rate
    '9279-1',    # respiratory rate
    '70274-6',   # MAP
    '2571-8',    # triglycerides
    '4544-3',    # HDL
    '2093-3',    # LDL
    '2085-9'     # total cholesterol
}

df_obs_features = df_obs_features[
    df_obs_features['code'].isin(feature_codes)
    ].copy()

df_obs_features.shape

df_obs_features.head()

df_obs_features.groupby('code')['patient_id'].nunique()

df_cohort.head()

# save df_cohort and df_obs_features to csv and transiiton to feature engineering
df_cohort.to_csv(
    '/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Cleaned/df_cohort.csv',
    index=False
),
df_obs_features.to_csv(
    '/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Cleaned/df_obs_features.csv',
    index=False
)
