==> Notebooks_Diabetes_Prediction/03_Feature_Engineering.ipynb <==
# Project: Predicting Diabetes Onset FHIR EHR Data
### Notebook: 03_Feature_Engineering.ipynb
### Purpose: Develop features for modeling

# import libraries
import pandas as pd
import numpy as np
from typing import List, Tuple


# load cohort data
df_cohort = pd.read_csv('/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Cleaned/df_cohort.csv',
                       parse_dates=['index_date', 'diabetes_onset_date', 'last_encounter_date'])

# load observations features df
df_obs_features = pd.read_csv('/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Cleaned/df_obs_features.csv',
                             parse_dates=['time', 'index_date'])

# inspect cohort df
df_cohort.head(), df_cohort.shape, df_cohort.dtypes

# inspect ibs features df 
df_obs_features.head(), df_obs_features.shape, df_obs_features.dtypes

# before merging data, check that all index 
# dates associated wiht patient_id are indentical in each dataframe
check = (
    df_obs_features[['patient_id','index_date']]
    .drop_duplicates()
    .merge(
        df_cohort[['patient_id','index_date']],
        on='patient_id',
        suffixes=('_obs','_cohort')
    )
)

# verify that both index date values for each row are identical match
(check['index_date_obs']!=check['index_date_cohort']).sum()

# drop index_date column from obs features df
df_obs_features = df_obs_features.drop(columns=['index_date'])
df_obs_features.head()

# merge obs features df and cohort df
df_features = df_obs_features.merge(
    df_cohort[['patient_id', 'index_date']],
    on='patient_id',
    how='inner'
)

df_features = df_features[
    (df_features['time'] < df_features['index_date']) &
    (df_features['time'] >= df_features['index_date'] - pd.Timedelta(days=365))
]

# Final verificaiton chekpoint to ensure index_dare column dropped and time column is
# a time before index_date ffrom cohort df 
assert 'index_date' not in df_obs_features.columns
assert (df_features['time'] < df_features['index_date']).all()

df_features.head()

# rename time columns to obs_time for clarity

df_features.rename(columns={'time':'obs_time'}, inplace=True)

# aggregate labs per patient
lab_agg = (
    df_features
    .groupby(['patient_id', 'code'])
    .agg(
        mean_value=('value', 'mean'),
        median_value=('value', 'median'),
        min_value=('value', 'min'),
        max_value=('value', 'max'),
        std_value=('value', 'std'),
        count_value=('value', 'count'),
        last_value=('value', 'last'),
        last_date=('obs_time', 'last'),
        index_date=('index_date', 'first')

    )
    .reset_index()
)

lab_agg.head()


# add a column days_since_last to get days between index date and last observation date
lab_agg['days_since_last'] = (
    lab_agg['index_date'] - lab_agg['last_date']
).dt.days

lab_agg.head()

# confirm no duplicate index_dates per patient

(
    lab_agg
    .groupby('patient_id')['index_date']
    .nunique()
    .max()
)

# map lab codes to lab names

lab_dict = {
    '8867-4': 'heart_rate',
    '9279-1': 'creatinine',
    '29463-7': 'body_weight',
    '8302-2': 'body_height',
    '39156-5': 'bmi',
    '70274-6': 'blood_glucose',
    '2085-9': 'hdl_cholesterol',
    '2093-3': 'total_cholesterol',
    '2571-8': 'ldl_cholesterol',
    '4544-3': 'hba1c'
}

# map to dicitonary
lab_agg['lab_name'] = lab_agg['code'].map(lab_dict)

# check for unmapped items
lab_agg[lab_agg['lab_name'].isna()]['code'].unique()


lab_agg.head()

# two labs have been incorrectly mapped, their coes
# correspond to generalized anxiety disorder and hematocrit
# they will be dropped

drop_codes = ['70274-6','4544-3']

lab_agg = lab_agg[~lab_agg['code'].isin(drop_codes)]

# Helper function to pivot columns

def pivot_features(df, index: str, columns: str, values: str) -> pd.DataFrame:
    return df.pivot(index = index,
                    columns = columns,
                    values = values
            )

# Helper function to add suffix to column names
def add_suffix(df, suffix: str) -> List[str]:
    names = []
    for c in df.columns:
        col_name = f"{c}_{suffix}"
        names.append(col_name)
    df.columns = names
    return df.columns

# pivot mean Values by lab name

mean_features = pivot_features(lab_agg, index='patient_id', columns='lab_name', values='mean_value')
mean_features.columns = add_suffix(mean_features, 'mean')
mean_features.head()

# pivot last value by lab names

last_value = pivot_features(lab_agg, index='patient_id', columns='lab_name', values='last_value')
last_value.columns = add_suffix(last_value, 'last_value')
last_value.head()

# pivot count values by lab names

count_features = pivot_features(lab_agg, index='patient_id', columns='lab_name', values='count_value')
count_features.columns = add_suffix(count_features,'count_value')
count_features.head()

# pivot recency features by lab name

recency_features = pivot_features(lab_agg, index='patient_id', columns='lab_name', values='days_since_last')
recency_features.columns = add_suffix(recency_features,'days_since_last')
recency_features.head()

# merge all feature dataframes

df_patient_features = (
    mean_features
    .merge(last_value, on='patient_id', how='left')
    .merge(count_features, on='patient_id', how='left')
    .merge(recency_features, on='patient_id', how='left')
)

# inspect patient features dataframe
df_patient_features.head()

# add missing flags 
missing_flags = df_patient_features.isna().astype(int)
missing_flags.columns = [c + '_missing' for c in missing_flags.columns]

df_patient_features = df_patient_features.merge(
    missing_flags,
    left_index=True,
    right_index=True
)

df_patient_features.head()

# utilization features table 

util_features = (
    df_features
    .groupby('patient_id')
    .agg(
        total_labs = ('code', 'count'),
        unique_labs = ('code', 'nunique')
    )
)

# merge utilization features into patient features dataframe

df_patient_features = df_patient_features.merge(
    util_features,
    left_index=True,
    right_index=True
)

# Merge labels for modeling
X = df_patient_features.merge(
    df_cohort[['patient_id','label']],
    on='patient_id',
    how='left'
)

X.columns

# remove redundant columns

cols_to_keep = [
    'patient_id', 'label',
    'bmi_mean', 'body_height_mean', 'body_weight_mean', 'creatinine_mean',
    'hdl_cholesterol_mean', 'heart_rate_mean', 'ldl_cholesterol_mean', 'total_cholesterol_mean',
    'bmi_last_value', 'body_height_last_value', 'body_weight_last_value', 'creatinine_last_value',
    'hdl_cholesterol_last_value', 'heart_rate_last_value', 'ldl_cholesterol_last_value', 'total_cholesterol_last_value',
    'bmi_count_value', 'body_height_count_value', 'body_weight_count_value', 'creatinine_count_value',
    'hdl_cholesterol_count_value', 'heart_rate_count_value', 'ldl_cholesterol_count_value', 'total_cholesterol_count_value',
    'bmi_days_since_last', 'body_height_days_since_last', 'body_weight_days_since_last', 'creatinine_days_since_last',
    'hdl_cholesterol_days_since_last', 'heart_rate_days_since_last', 'ldl_cholesterol_days_since_last', 'total_cholesterol_days_since_last',
    'creatinine_mean_missing', 'hdl_cholesterol_mean_missing', 'ldl_cholesterol_mean_missing', 'total_cholesterol_mean_missing',
    'total_labs', 'unique_labs'
]

df_features_master  = X[cols_to_keep].copy()

# inspect master features dataset

df_features_master.head()

# save master features dataframe

df_features_master.to_csv('/Users/sanasiddiqui/Desktop/Current Desktop/Data Portfolio Projects/Diabetes_Onset_Prediction_FHIR/Data_Diabetes_Prediction/Cleaned/df_features_master.csv', index=False)
